add connections api
