+++ viewer/viewer.js
@@ -72,6 +72,7 @@ app.getpost = (route, mw, func) => { app.get(route, mw, func); app.post(route, m
 let { internals } = require('./internals')(Config, RE2, http, https);
 let ViewerUtils = require('./viewerUtils')(async, Db, Config, molochparser, internals);
 let sessionAPIs = require('./apiSessions')(async, util, Pcap, Db, Config, ViewerUtils, molochparser, internals);
+let connectionAPIs = require('./apiConnections')(async, Db, Config, ViewerUtils, sessionAPIs);
 
 function isProduction () {
   return app.get('env') === 'production';
@@ -433,26 +434,13 @@ process.on('SIGINT', function () {
   process.exit();
 });
 
-function loadFields () {
-  Db.loadFields(function (err, data) {
-    if (err) { data = []; } else { data = data.hits.hits; }
-
-    // Everything will use dbField2 as dbField
-    for (let i = 0, ilen = data.length; i < ilen; i++) {
-      internals.oldDBFields[data[i]._source.dbField] = data[i]._source;
-      data[i]._source.dbField = data[i]._source.dbField2;
-      if (data[i]._source.portField2) {
-        data[i]._source.portField = data[i]._source.portField2;
-      } else {
-        delete data[i]._source.portField;
-      }
-      delete data[i]._source.rawField;
-    }
-    Config.loadFields(data);
-    app.locals.fieldsMap = JSON.stringify(Config.getFieldsMap());
-    app.locals.fieldsArr = Config.getFields().sort(function (a, b) { return (a.exp > b.exp ? 1 : -1); });
-    createSessionDetail();
-  });
+function setFieldLocals () {
+  ViewerUtils.loadFields()
+    .then((result) => {
+      app.locals.fieldsMap = result.fieldsMap;
+      app.locals.fieldsArr = result.fieldsArr;
+      createSessionDetail();
+    });
 }
 
 function loadPlugins () {
@@ -677,7 +665,6 @@ class Mutex {
 // ----------------------------------------------------------------------------
 // Requests
 // ----------------------------------------------------------------------------
-
 function addAuth (info, user, node, secret) {
   if (!info.headers) {
     info.headers = {};
@@ -705,14 +692,6 @@ function addCaTrust (info, node) {
   if (internals.caTrustCerts[node] !== undefined && internals.caTrustCerts[node].length > 0) {
     info.ca = internals.caTrustCerts[node];
     info.agent.options.ca = internals.caTrustCerts[node];
-  }
-}
-
-function noCache (req, res, ct) {
-  res.header('Cache-Control', 'no-cache, private, no-store, must-revalidate, max-stale=0, post-check=0, pre-check=0');
-  if (ct) {
-    res.setHeader('Content-Type', ct);
-    res.header('X-Content-Type-Options', 'nosniff');
   }
 }
 
@@ -748,7 +727,7 @@ function getViewUrl (node, cb) {
 }
 
 function proxyRequest (req, res, errCb) {
-  noCache(req, res);
+  ViewerUtils.noCache(req, res);
 
   getViewUrl(req.params.nodeName, function (err, viewUrl, client) {
     if (err) {
@@ -4064,6 +4043,7 @@ app.get('/:nodeName/:fileNum/filesize.json', [noCacheJson, checkPermissions(['hi
   });
 });
 
+// TODO ECR - put in sessions?
 app.use('/buildQuery.json', [noCacheJson, logAction('query')], function (req, res, next) {
   if (req.method === 'POST') {
     req.query = req.body;
@@ -4088,37 +4068,9 @@ app.use('/buildQuery.json', [noCacheJson, logAction('query')], function (req, re
   });
 });
 
-// TODO ECR - document
-app.getpost(
-  ['/api/sessions', '/sessions.json'],
-  [noCacheJson, recordResponseTime, logAction('sessions'), setCookie, fillBody],
-  sessionAPIs.getSessions
-);
-
-// TODO ECR - document
-app.getpost(
-  ['/api/spiview', '/spiview.json'],
-  [noCacheJson, recordResponseTime, logAction('spiview'), setCookie, fillBody],
-  sessionAPIs.getSPIView
-);
-
-// TODO ECR - document
-app.getpost(
-  ['/api/spigraph', '/spigraph.json'],
-  [noCacheJson, recordResponseTime, logAction('spigraph'), setCookie, fillBody, fieldToExp],
-  sessionAPIs.getSPIGraph
-);
-
-// TODO ECR
-app.getpost(
-  ['/api/spigraphhierarchy', 'spigraphhierarchy'],
-  [noCacheJson, recordResponseTime, logAction('spigraph'), setCookie, fillBody],
-  sessionAPIs.getSPIGraphHierarchy
-);
-
-app.get('/reverseDNS.txt', [noCacheJson, logAction()], function (req, res) {
-  // console.log('reverseDNS.txt', req.query);
-  dns.reverse(req.query.ip, function (err, data) {
+// TODO ECR - put this elsewhere?
+app.get('/reverseDNS.txt', [noCacheJson, logAction()], (req, res) => {
+  dns.reverse(req.query.ip, (err, data) => {
     if (err) {
       return res.send('reverse error');
     }
@@ -4126,470 +4078,44 @@ app.get('/reverseDNS.txt', [noCacheJson, logAction()], function (req, res) {
   });
 });
 
-// ----------------------------------------------------------------------------
-// buildConnectionQuery(req, fields, options, fsrc, fdst, dstipport, resultId, cb)
-//
-// Returns (via "return cb(...)") an array of 1..2 connection query objects
-// (see the definition of "result" at the beginning of the function), depending on
-// whether or not baseline is enabled. The query and indices are initially returned
-// from buildSessionQuery and then adjusted by this function.
-//
-// The queries represented by these objects can be executed via
-// dbConnectionQuerySearch.
-//
-// This code was factored out from buildConnections.
-// ----------------------------------------------------------------------------
-function buildConnectionQuery (req, fields, options, fsrc, fdst, dstipport, resultId, cb) {
-  let result = {
-    resultId: resultId,
-    err: null,
-    query: null,
-    indices: null,
-    options: options
-  };
+// Sessions Endpoints ---------------------------------------------------------
+app.getpost( // sessions endpoint (POST or GET)
+  ['/api/sessions', '/sessions.json'],
+  [noCacheJson, recordResponseTime, logAction('sessions'), setCookie, fillBody],
+  sessionAPIs.getSessions
+);
 
-  // If network graph baseline is enabled (enabled: req.query.baselineDate != 0, disabled:req.query.baselineDate=0 or undefined)
-  //   then two queries will be run (ie., run buildSessionQuery->searchPrimary->process twice): first for the
-  //   original specified time frame and second for the same time frame immediately preceding it.
-  //
-  // Nodes have an .inresult attribute where:
-  //   0 = 00 = not in either result set (although you'll never see these, obviously)
-  //   1 = 01 = seen during the "current" time frame but not in the "baseline" time frame (ie., "new")
-  //   2 = 10 = seen during the "baseline" time frame but not in the "current" time frame (ie., "old")
-  //   3 = 11 = seen during both the "current" time frame and the "baseline" time frame
-  // This is only performed where startTime/startTime are defined, and never for "all" time range (date=-1).
-  //
-  // With baselining, req.query.baselineDate can determine baseline time frame, which is the number of
-  // hours prior to the "start" query time, similar to req.query.date. If unspecified or zero, baseline
-  // uses the immediate time frame of the same duration immediately prior to the req.query.startTime.
-  // However, If req.query.baselineDate ends with x, the duration of the baseline is the time frame of
-  // the "current" time frame multiplied by that number.
-  let doBaseline = false;
-  let baselineDate = 0;
-  let baselineDateIsMultiplier = false;
+app.getpost( // spiview endpoint (POST or GET)
+  ['/api/spiview', '/spiview.json'],
+  [noCacheJson, recordResponseTime, logAction('spiview'), setCookie, fillBody],
+  sessionAPIs.getSPIView
+);
 
-  if (((req.query.baselineDate !== undefined) && (req.query.baselineDate.length !== 0) && (String(req.query.baselineDate) !== '0') &&
-        (req.query.date !== '-1') && (req.query.startTime !== undefined) && (req.query.stopTime !== undefined)) ||
-      (resultId > 1)) {
-    doBaseline = true;
-  }
+app.getpost( // spigraph endpoint (POST or GET)
+  ['/api/spigraph', '/spigraph.json'],
+  [noCacheJson, recordResponseTime, logAction('spigraph'), setCookie, fillBody, fieldToExp],
+  sessionAPIs.getSPIGraph
+);
 
-  if (doBaseline) {
-    let baselineDateTmpStr = req.query.baselineDate;
-    if (baselineDateTmpStr.endsWith('x')) {
-      baselineDateIsMultiplier = true;
-      baselineDateTmpStr = baselineDateTmpStr.slice(0, -1);
-    }
-    baselineDate = parseInt(baselineDateTmpStr, 10);
-    doBaseline = (doBaseline && (baselineDate > 0));
-    baselineDateIsMultiplier = (doBaseline && baselineDateIsMultiplier && (baselineDate > 0));
-  }
+app.getpost( // spigraph hierarchy endpoint (POST or GET)
+  ['/api/spigraphhierarchy', 'spigraphhierarchy'],
+  [noCacheJson, recordResponseTime, logAction('spigraph'), setCookie, fillBody],
+  sessionAPIs.getSPIGraphHierarchy
+);
 
-  // use a copy of req.query as we will modify the startTime/stopTime if we are doing a baseline query
-  let tmpReqQuery = JSON.parse(JSON.stringify(req.query));
+// TODO ECR
+// Connections Endpoints ------------------------------------------------------
+app.getpost( // connections endpoint (POST or GET)
+  ['/api/connections', '/connections.json'],
+  [noCacheJson, recordResponseTime, logAction('connections'), setCookie, fillBody],
+  connectionAPIs.getConnections
+);
 
-  if (resultId > 1) {
-    // replace current time frame start/stop values with baseline time frame start/stop values
-    let currentQueryTimes = ViewerUtils.determineQueryTimes(req.query);
-    if (Config.debug) {
-      console.log('buildConnections baseline.0', 'startTime', currentQueryTimes[0], 'stopTime', currentQueryTimes[1], baselineDate, baselineDateIsMultiplier ? 'x' : '');
-    }
-    if ((currentQueryTimes[0] !== undefined) && (currentQueryTimes[1] !== undefined)) {
-      // baseline stop time ends 1 second prior to "current" start time
-      tmpReqQuery.stopTime = currentQueryTimes[0] - 1;
-      if ((baselineDate > 0) && (!baselineDateIsMultiplier)) {
-        // baseline time duration was specified (hours)
-        tmpReqQuery.startTime = tmpReqQuery.stopTime - (60 * 60 * baselineDate);
-      } else {
-        // baseline time frame is unspecified, so use the immediate prior time frame of same (or multiplied) duration
-        tmpReqQuery.startTime = tmpReqQuery.stopTime - ((currentQueryTimes[1] - currentQueryTimes[0]) * (baselineDateIsMultiplier ? baselineDate : 1));
-      }
-      if (Config.debug) {
-        console.log('buildConnections baseline.1', 'startTime', tmpReqQuery.startTime, 'stopTime', tmpReqQuery.stopTime, 'diff', (tmpReqQuery.stopTime - tmpReqQuery.startTime));
-      }
-    }
-  } // resultId > 1 (calculating baseline query time frame)
-
-  sessionAPIs.buildSessionQuery(req, (bsqErr, query, indices) => {
-    if (bsqErr) {
-      console.log('ERROR - buildConnectionQuery -> buildSessionQuery', resultId, bsqErr);
-      result.err = bsqErr;
-      return cb([result]);
-    } else {
-      query.query.bool.filter.push({ exists: { field: req.query.srcField } });
-      query.query.bool.filter.push({ exists: { field: req.query.dstField } });
-
-      query._source = fields;
-      query.docvalue_fields = [fsrc, fdst];
-
-      if (dstipport) {
-        query._source.push('dstPort');
-      }
-
-      result.query = JSON.parse(JSON.stringify(query));
-      result.indices = JSON.parse(JSON.stringify(indices));
-
-      if ((resultId === 1) && (doBaseline)) {
-        buildConnectionQuery(req, fields, options, fsrc, fdst, dstipport, resultId + 1, function (baselineResult) {
-          return cb([result].concat(baselineResult));
-        });
-      } else {
-        return cb([result]);
-      }
-    } // bsqErr if/else
-  }, tmpReqQuery); // buildSessionQuery
-} // buildConnectionQuery
-
-// ----------------------------------------------------------------------------
-// dbConnectionQuerySearch(connQueries, resultId, cb)
-//
-// Executes the query/queries specified in the connQueries array (elements are
-// of the type returned by buildConnectionQuery) by calling Db.searchPrimary
-// and returns the results via callback (see the definition of the "resultSet"
-// object at the beginning of this function). The results are returned in an
-// array containing the result sets which correspond to the queries in the
-// connQueries array.
-//
-// This code was factored out from buildConnections.
-// ----------------------------------------------------------------------------
-function dbConnectionQuerySearch (connQueries, cb) {
-  let resultSet = {
-    resultId: null,
-    err: null,
-    graph: null
-  };
-
-  if (connQueries.length > 0) {
-    resultSet.resultId = connQueries[0] ? connQueries[0].resultId : null;
-    resultSet.err = connQueries[0] ? connQueries[0].err : 'null query object';
-
-    if (((connQueries[0]) && (connQueries[0].err)) || (!connQueries[0])) {
-      // propogate query errors up into the result set without doing a search
-      console.log('ERROR - buildConnectionQuery -> dbConnectionQuerySearch', resultSet.resultId, resultSet.err);
-      return cb([resultSet]);
-    } else {
-      Db.searchPrimary(connQueries[0].indices, 'session', connQueries[0].query, connQueries[0].options, function (err, graph) {
-        if (err || graph.error) {
-          console.log('ERROR - dbConnectionQuerySearch -> Db.searchPrimary', connQueries[0].resultId, err, graph.error);
-          resultSet.err = err || graph.error;
-        }
-        resultSet.graph = graph;
-        if (connQueries.length > 1) {
-          dbConnectionQuerySearch(connQueries.slice(1), function (baselineResultSet) {
-            return cb([resultSet].concat(baselineResultSet));
-          });
-        } else {
-          return cb([resultSet]);
-        }
-      }); // Db.searchPrimary
-    } // if connQueries[0].err) / else
-  } else {
-    return cb([null]);
-  } // (connQueries.length > 0) / else
-} // dbConnectionQuerySearch
-
-// ----------------------------------------------------------------------------
-// buildConnections(req, res, cb)
-//
-// Returns objects needed to populate the graph of logical connections between
-// nodes representing fields of sessions.
-//
-// function flow is:
-//
-// 0. buildConnections
-// 1. buildConnectionQuery       - creates array of 1..2 connQueries
-// 2. dbConnectionQuerySearch    - executes connQueries searches via Db.searchPrimary
-// 3. processResultSets          - accumulate nodes and links into nodesHash/connects hashes
-//    - process
-//      - updateValues
-// 4. processResultSets callback - distill nodesHash/connects hashes into
-//                                 nodes/links arrays and return
-// ----------------------------------------------------------------------------
-function buildConnections (req, res, cb) {
-  let dstipport;
-  if (req.query.dstField === 'ip.dst:port') {
-    dstipport = true;
-    req.query.dstField = 'dstIp';
-  }
-
-  req.query.srcField = req.query.srcField || 'srcIp';
-  req.query.dstField = req.query.dstField || 'dstIp';
-  let fsrc = req.query.srcField;
-  let fdst = req.query.dstField;
-  let minConn = req.query.minConn || 1;
-
-  // get the requested fields
-  let fields = ['totBytes', 'totDataBytes', 'totPackets', 'node'];
-  if (req.query.fields) { fields = req.query.fields.split(','); }
-
-  let options;
-  if (req.query.cancelId) { options = { cancelId: `${req.user.userId}::${req.query.cancelId}` }; }
-
-  let dstIsIp = fdst.match(/(\.ip|Ip)$/);
-
-  let nodesHash = {};
-  let connects = {};
-  let nodes = [];
-  let links = [];
-  let totalHits = 0;
-
-  // ----------------------------------------------------------------------------/
-  // updateValues and process are for aggregating query results into their final form
-  let dbFieldsMap = Config.getDBFieldsMap();
-  function updateValues (data, property, fields) {
-    for (let i in fields) {
-      let dbField = fields[i];
-      let field = dbFieldsMap[dbField];
-      if (data.hasOwnProperty(dbField)) {
-        // sum integers
-        if (field.type === 'integer' && field.category !== 'port') {
-          property[dbField] = (property[dbField] || 0) + data[dbField];
-        } else { // make a list of values
-          if (!property[dbField]) { property[dbField] = []; }
-          // make all values an array (because sometimes they are by default)
-          let values = [ data[dbField] ];
-          if (Array.isArray(data[dbField])) {
-            values = data[dbField];
-          }
-          for (let value of values) {
-            property[dbField].push(value);
-          }
-          if (property[dbField] && Array.isArray(property[dbField])) {
-            property[dbField] = [ ...new Set(property[dbField]) ]; // unique only
-          }
-        }
-      }
-    }
-  } // updateValues
-
-  // ----------------------------------------------------------------------------/
-  function process (vsrc, vdst, f, fields, resultId) {
-    // ES 6 is returning formatted timestamps instead of ms like pre 6 did
-    // https://github.com/elastic/elasticsearch/issues/27740
-    if (vsrc.length === 24 && vsrc[23] === 'Z' && vsrc.match(/^\d\d\d\d-\d\d-\d\dT\d\d:\d\d:\d\d.\d\d\dZ$/)) {
-      vsrc = new Date(vsrc).getTime();
-    }
-    if (vdst.length === 24 && vdst[23] === 'Z' && vdst.match(/^\d\d\d\d-\d\d-\d\dT\d\d:\d\d:\d\d.\d\d\dZ$/)) {
-      vdst = new Date(vdst).getTime();
-    }
-
-    if (nodesHash[vsrc] === undefined) {
-      nodesHash[vsrc] = { id: `${vsrc}`, cnt: 0, sessions: 0, inresult: 0 };
-    }
-
-    nodesHash[vsrc].sessions++;
-    nodesHash[vsrc].type |= 1;
-    nodesHash[vsrc].inresult |= resultId;
-    updateValues(f, nodesHash[vsrc], fields);
-
-    if (nodesHash[vdst] === undefined) {
-      nodesHash[vdst] = { id: `${vdst}`, cnt: 0, sessions: 0, inresult: 0 };
-    }
-
-    nodesHash[vdst].sessions++;
-    nodesHash[vdst].type |= 2;
-    nodesHash[vdst].inresult |= resultId;
-    updateValues(f, nodesHash[vdst], fields);
-
-    let linkId = `${vsrc}->${vdst}`;
-    if (connects[linkId] === undefined) {
-      connects[linkId] = { value: 0, source: vsrc, target: vdst };
-      nodesHash[vsrc].cnt++;
-      nodesHash[vdst].cnt++;
-    }
-
-    connects[linkId].value++;
-    updateValues(f, connects[linkId], fields);
-  } // process
-
-  // ----------------------------------------------------------------------------//////////
-  // processResultSets - process the hits of each search resultset into nodesHash and connects
-  function processResultSets (connResultSets, cb) {
-    let resultSetStatus = {
-      resultId: null,
-      err: null,
-      hits: 0
-    };
-
-    if (connResultSets.length > 0) {
-      resultSetStatus.resultId = connResultSets[0] ? connResultSets[0].resultId : null;
-      resultSetStatus.err = connResultSets[0] ? connResultSets[0].err : 'null resultset';
-
-      if (((connResultSets[0]) && (connResultSets[0].err)) || (!connResultSets[0])) {
-        // propogate errors up (and stop processing)
-        console.log('ERROR - buildConnectionQuery -> processResultSets', resultSetStatus.resultId, resultSetStatus.err);
-        return cb([resultSetStatus]);
-      } else {
-        async.eachLimit(connResultSets[0].graph.hits.hits, 10, function (hit, hitCb) {
-          let f = hit._source;
-          f = ViewerUtils.flattenFields(f);
-
-          let asrc = hit.fields[fsrc];
-          let adst = hit.fields[fdst];
-
-          if (asrc === undefined || adst === undefined) {
-            return setImmediate(hitCb);
-          }
-
-          if (!Array.isArray(asrc)) { asrc = [asrc]; }
-          if (!Array.isArray(adst)) { adst = [adst]; }
-
-          for (let vsrc of asrc) {
-            for (let vdst of adst) {
-              if (dstIsIp && dstipport) {
-                if (vdst.includes(':')) {
-                  vdst += '.' + f.dstPort;
-                } else {
-                  vdst += ':' + f.dstPort;
-                }
-              }
-              process(vsrc, vdst, f, fields, connResultSets[0].resultId);
-            } // let vdst of adst
-          } // for vsrc of asrc
-          setImmediate(hitCb);
-        }, function (err) {
-          resultSetStatus.err = err;
-          resultSetStatus.hits = connResultSets[0].graph.hits.total;
-          if (connResultSets.length > 1) {
-            processResultSets(connResultSets.slice(1), function (baselineResultSetStatus) {
-              return cb([resultSetStatus].concat(baselineResultSetStatus));
-            });
-          } else {
-            return cb([resultSetStatus]);
-          }
-        }); // async.eachLimit(graph.hits.hits) / function(err)
-      } // if connResultSets[0].err) / else
-    } else {
-      return cb([null]);
-    } // (connResultSets.length > 0) / else
-  } // processResultSets
-
-  // ----------------------------------------------------------------------------//////////
-  // call to build the session query|queries and indices
-  buildConnectionQuery(req, fields, options, fsrc, fdst, dstipport, 1, function (connQueries) {
-    if (Config.debug) {
-      console.log('buildConnections.connQueries', connQueries.length, JSON.stringify(connQueries, null, 2));
-    }
-
-    // ONE or TWO session queries will be executed, depending on if baseline is enabled:
-    //   1. for the "current" time frame, the one specified originally in req.query
-    //   2. for the "baseline" time frame immediately prior to the time frame of "1."
-    //      (only if baseline is enabled)
-    // The call to process() will ensure the resultId value is OR'ed into the .inresult
-    //   attribute of each node.
-
-    // prepare and execute the Db.searchPrimary query|queries
-
-    if (connQueries.length > 0) {
-      dbConnectionQuerySearch(connQueries, function (connResultSets) {
-        if (Config.debug) {
-          console.log('buildConnections.connResultSets', connResultSets.length, JSON.stringify(connResultSets, null, 2));
-        }
-
-        // aggregate final return values for nodes and links
-        processResultSets(connResultSets, function (connResultSetStats) {
-          if (Config.debug) {
-            console.log('buildConnections.processResultSets', connResultSetStats.length, JSON.stringify(connResultSetStats, null, 2));
-          }
-
-          for (let stat of connResultSetStats) {
-            if (stat.err) {
-              return cb(stat.err, null, null, null);
-            }
-            totalHits += stat.hits;
-          }
-
-          let nodeKeys = Object.keys(nodesHash);
-          if (Config.get('regressionTests', false)) {
-            nodeKeys = nodeKeys.sort(function (a, b) { return nodesHash[a].id.localeCompare(nodesHash[b].id); });
-          }
-          for (let node of nodeKeys) {
-            if (nodesHash[node].cnt < minConn) {
-              nodesHash[node].pos = -1;
-            } else {
-              nodesHash[node].pos = nodes.length;
-              nodes.push(nodesHash[node]);
-            }
-          }
-
-          for (let key in connects) {
-            var c = connects[key];
-            c.source = nodesHash[c.source].pos;
-            c.target = nodesHash[c.target].pos;
-            if (c.source >= 0 && c.target >= 0) {
-              links.push(connects[key]);
-            }
-          }
-
-          if (Config.debug) {
-            console.log('buildConnections.nodesHash', nodesHash);
-            console.log('buildConnections.connects', connects);
-            console.log('buildConnections.nodes', nodes.length, nodes);
-            console.log('buildConnections.links', links.length, links);
-            console.log('buildConnections.totalHits', totalHits);
-          }
-
-          return cb(null, nodes, links, totalHits);
-        }); // processResultSets.callback
-      }); // dbConnectionQuerySearch.callback
-    } else {
-      let err = 'no connection queries generated';
-      console.log('ERROR - buildConnections', err);
-      return cb(err, null, null, null);
-    } // connQueries.length check
-  }); // buildConnectionQuery.callback
-} // buildConnections
-
-app.get('/connections.json', [noCacheJson, recordResponseTime, logAction('connections'), setCookie], (req, res) => {
-  let health;
-  Db.healthCache(function (err, h) { health = h; });
-  buildConnections(req, res, function (err, nodes, links, total) {
-    if (err) { return res.molochError(403, err.toString()); }
-    res.send({ health: health, nodes: nodes, links: links, recordsFiltered: total });
-  });
-});
-
-app.get('/connections.csv', logAction(), function (req, res) {
-  noCache(req, res, 'text/csv');
-
-  var seperator = req.query.seperator || ',';
-  buildConnections(req, res, function (err, nodes, links, total) {
-    if (err) {
-      return res.send(err);
-    }
-
-    // write out the fields requested
-    let fields = ['totBytes', 'totDataBytes', 'totPackets', 'node'];
-    if (req.query.fields) { fields = req.query.fields.split(','); }
-
-    res.write('Source, Destination, Sessions');
-    let displayFields = {};
-    for (let field of fields) {
-      let fieldsMap = JSON.parse(app.locals.fieldsMap);
-      for (let f in fieldsMap) {
-        if (fieldsMap[f].dbField === field) {
-          let friendlyName = fieldsMap[f].friendlyName;
-          displayFields[field] = fieldsMap[f];
-          res.write(`, ${friendlyName}`);
-        }
-      }
-    }
-    res.write('\r\n');
-
-    for (let i = 0, ilen = links.length; i < ilen; i++) {
-      res.write('"' + nodes[links[i].source].id.replace('"', '""') + '"' + seperator +
-                '"' + nodes[links[i].target].id.replace('"', '""') + '"' + seperator +
-                     links[i].value + seperator);
-      for (let f = 0, flen = fields.length; f < flen; f++) {
-        res.write(links[i][displayFields[fields[f]].dbField].toString());
-        if (f !== flen - 1) { res.write(seperator); }
-      }
-      res.write('\r\n');
-    }
-
-    res.end();
-  });
-});
+app.getpost( // connections csv endpoint (POST or GET)
+  ['/api/connections.csv', '/connections.csv'],
+  [logAction('connections.csv'), fillBody],
+  connectionAPIs.getConnectionsCSV
+);
 
 function csvListWriter (req, res, list, fields, pcapWriter, extension) {
   if (list.length > 0 && list[0].fields) {
@@ -4647,8 +4173,9 @@ function csvListWriter (req, res, list, fields, pcapWriter, extension) {
   res.end();
 }
 
+// TODO ECR
 app.get(/\/sessions.csv.*/, logAction(), function (req, res) {
-  noCache(req, res, 'text/csv');
+  ViewerUtils.noCache(req, res, 'text/csv');
 
   // default fields to display in csv
   var fields = ['ipProtocol', 'firstPacket', 'lastPacket', 'srcIp', 'srcPort', 'srcGEO', 'dstIp', 'dstPort', 'dstGEO', 'totBytes', 'totDataBytes', 'totPackets', 'node'];
@@ -4673,7 +4200,7 @@ app.get(/\/sessions.csv.*/, logAction(), function (req, res) {
 });
 
 app.get('/multiunique.txt', logAction(), function (req, res) {
-  noCache(req, res, 'text/plain; charset=utf-8');
+  ViewerUtils.noCache(req, res, 'text/plain; charset=utf-8');
 
   if (req.query.exp === undefined) {
     return res.send('Missing exp parameter');
@@ -4753,7 +4280,7 @@ app.get('/multiunique.txt', logAction(), function (req, res) {
 });
 
 app.get('/unique.txt', [logAction(), fillBody, fieldToExp], function (req, res) {
-  noCache(req, res, 'text/plain; charset=utf-8');
+  ViewerUtils.noCache(req, res, 'text/plain; charset=utf-8');
 
   if (req.query.field === undefined && req.query.exp === undefined) {
     return res.send('Missing field or exp parameter');
@@ -5359,7 +4886,7 @@ app.get('/:nodeName/session/:id/detail', cspHeader, logAction(), (req, res) => {
  */
 app.get('/:nodeName/session/:id/packets', [logAction(), checkPermissions(['hidePcap'])], (req, res) => {
   isLocalView(req.params.nodeName, function () {
-    noCache(req, res);
+    ViewerUtils.noCache(req, res);
     req.packetsOnly = true;
     localSessionDetail(req, res);
   },
@@ -5490,7 +5017,7 @@ app.get('/bodyHash/:hash', logAction('bodyhash'), function (req, res) {
                 res.status(400);
                 return res.end(err);
               } else if (item) {
-                noCache(req, res, 'application/force-download');
+                ViewerUtils.noCache(req, res, 'application/force-download');
                 res.setHeader('content-disposition', contentDisposition(item.bodyName + '.pellet'));
                 return res.end(item.data);
               } else {
@@ -5522,7 +5049,7 @@ app.get('/:nodeName/:id/bodyHash/:hash', checkProxyRequest, function (req, res) 
       res.status(400);
       return res.end(err);
     } else if (item) {
-      noCache(req, res, 'application/force-download');
+      ViewerUtils.noCache(req, res, 'application/force-download');
       res.setHeader('content-disposition', contentDisposition(item.bodyName + '.pellet'));
       return res.end(item.data);
     } else {
@@ -5683,14 +5210,14 @@ function writePcapNg (res, id, options, doneCb) {
 }
 
 app.get('/:nodeName/pcapng/:id.pcapng', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], (req, res) => {
-  noCache(req, res, 'application/vnd.tcpdump.pcap');
+  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
   writePcapNg(res, req.params.id, { writeHeader: !req.query || !req.query.noHeader || req.query.noHeader !== 'true' }, function () {
     res.end();
   });
 });
 
 app.get('/:nodeName/pcap/:id.pcap', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], (req, res) => {
-  noCache(req, res, 'application/vnd.tcpdump.pcap');
+  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
 
   writePcap(res, req.params.id, { writeHeader: !req.query || !req.query.noHeader || req.query.noHeader !== 'true' }, function () {
     res.end();
@@ -5698,7 +5225,7 @@ app.get('/:nodeName/pcap/:id.pcap', [checkProxyRequest, checkPermissions(['disab
 });
 
 app.get('/:nodeName/raw/:id.png', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], function (req, res) {
-  noCache(req, res, 'image/png');
+  ViewerUtils.noCache(req, res, 'image/png');
 
   processSessionIdAndDecode(req.params.id, 1000, function (err, session, results) {
     if (err) {
@@ -5729,7 +5256,7 @@ app.get('/:nodeName/raw/:id.png', [checkProxyRequest, checkPermissions(['disable
 });
 
 app.get('/:nodeName/raw/:id', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], function (req, res) {
-  noCache(req, res, 'application/vnd.tcpdump.pcap');
+  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
 
   processSessionIdAndDecode(req.params.id, 10000, function (err, session, results) {
     if (err) {
@@ -5743,7 +5270,7 @@ app.get('/:nodeName/raw/:id', [checkProxyRequest, checkPermissions(['disablePcap
 });
 
 app.get('/:nodeName/entirePcap/:id.pcap', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], (req, res) => {
-  noCache(req, res, 'application/vnd.tcpdump.pcap');
+  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
 
   var options = { writeHeader: true };
 
@@ -5824,7 +5351,7 @@ function sessionsPcapList (req, res, list, pcapWriter, extension) {
 }
 
 function sessionsPcap (req, res, pcapWriter, extension) {
-  noCache(req, res, 'application/vnd.tcpdump.pcap');
+  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
 
   if (req.query.ids) {
     var ids = ViewerUtils.queryValueToArray(req.query.ids);
@@ -7514,7 +7041,7 @@ function pcapScrub (req, res, sid, whatToRemove, endCb) {
 }
 
 app.get('/:nodeName/delete/:whatToRemove/:sid', [checkProxyRequest, checkPermissions(['removeEnabled'])], (req, res) => {
-  noCache(req, res);
+  ViewerUtils.noCache(req, res);
 
   res.statusCode = 200;
 
@@ -7690,7 +7217,7 @@ function sendSessionWorker (options, cb) {
 internals.sendSessionQueue = async.queue(sendSessionWorker, 10);
 
 app.get('/:nodeName/sendSession/:id', checkProxyRequest, function (req, res) {
-  noCache(req, res);
+  ViewerUtils.noCache(req, res);
   res.statusCode = 200;
 
   var options = {
@@ -7708,7 +7235,7 @@ app.get('/:nodeName/sendSession/:id', checkProxyRequest, function (req, res) {
 });
 
 app.post('/:nodeName/sendSessions', checkProxyRequest, function (req, res) {
-  noCache(req, res);
+  ViewerUtils.noCache(req, res);
   res.statusCode = 200;
 
   if (req.body.ids === undefined ||
@@ -8473,8 +8000,8 @@ function main () {
     internals.previousNodesStats.push(info.nodes);
   });
 
-  loadFields();
-  setInterval(loadFields, 2 * 60 * 1000);
+  setFieldLocals();
+  setInterval(setFieldLocals, 2 * 60 * 1000);
 
   loadPlugins();
